refactor image_api to remove dependecy getting scooped into py2 tests
